Execute.c: code for implementing coordinate variable read/write with subscripting 	were added.
Machine.c OpsList.h Translate.c : New opcodes where inserted to support the special
	case of when one variable is assigned or a subsection of one
	variable is assigned to another, yet undefined, variable. In this situation
	all of the attributes, dimension info, and coordinates need to propagate
	to the lhs. The same is true for coordinate variable assignment and parameters.

defs.h: had a dim_num field in NclSelectionRecord that needed to be a long.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@1765 10fe2890-8a40-4696-8b1d-7ccbb19c398e

+++ ni/src/ncl/defs.h
@@ -1,6 +1,6 @@
 
 /*
- *      $Id: defs.h,v 1.8 1994-01-21 02:49:18 ethan Exp $
+ *      $Id: defs.h,v 1.9 1994-01-25 00:24:17 ethan Exp $
  */
 /************************************************************************
 *									*
@@ -164,7 +164,7 @@ typedef enum {
 
 typedef struct _NclSelection{
         NclSelectionTypes sel_type;
-	int dim_num;
+	long dim_num;
         union {
                 struct _NclSubscriptSelection sub;
                 struct _NclVectorSelection  vec;
